// src/generated/models/Mserp_purchaserequisitionheaderv2entitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mserp_purchaserequisitionheaderv2entitiesmserp_ispurchaserequisitiononhold = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionheaderv2entitiesmserp_ispurchaserequisitiononhold = keyof typeof Mserp_purchaserequisitionheaderv2entitiesmserp_ispurchaserequisitiononhold;
export const Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionpurpose = {
  200000000: 'Consumption',
  200000001: 'Replenishment'
} as const;
export type Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionpurpose = keyof typeof Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionpurpose;
export const Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionstatus = {
  200000000: 'Draft',
  200000001: 'Inreview',
  200000002: 'Rejected',
  200000003: 'Approved',
  200000004: 'Cancelled',
  200000005: 'Closed',
  200000006: 'Budgetreserved'
} as const;
export type Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionstatus = keyof typeof Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionstatus;

export interface Mserp_purchaserequisitionheaderv2entities {
  mserp_defaultaccountingdate?: string;
  mserp_defaultbusinessjustificationcode?: string;
  mserp_defaultbusinessjustificationdetails?: string;
  mserp_defaultprojectid?: string;
  mserp_defaultrequesteddate?: string;
  mserp_ispurchaserequisitiononhold?: Mserp_purchaserequisitionheaderv2entitiesmserp_ispurchaserequisitiononhold;
  mserp_ispurchaserequisitiononholdname?: string;
  mserp_onholdexplanation?: string;
  mserp_preparerpersonnelnumber?: string;
  mserp_projectbuyinglegalentityid?: string;
  mserp_purchaserequisitionheaderv2entityid: string;
  mserp_requisitionname?: string;
  mserp_requisitionnumber?: string;
  mserp_requisitionpurpose?: Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionpurpose;
  mserp_requisitionpurposename?: string;
  mserp_requisitionstatus?: Mserp_purchaserequisitionheaderv2entitiesmserp_requisitionstatus;
  mserp_requisitionstatusname?: string;
}