// src/generated/models/Mserp_purchaserequisitionlinev2entitiesModel.ts
export const Mserp_purchaserequisitionlinev2entitiesmserp_deliveryaddresstimezone = {
  200000000: '_GMT_12_00_InternationalDateLineWest',
  200000001: '_GMT_11_00_CoordinatedUniversalTime_11',
  200000002: '_GMT_10_00_Hawaii',
  200000003: '_GMT_09_00_Alaska',
  200000004: '_GMT_08_00_PacificTime_US_Canada_',
  200000005: '_GMT_08_00_Tijuana_BajaCalifornia',
  200000006: '_GMT_07_00_Arizona',
  200000007: '_GMT_07_00_MountainTime_US_Canada_',
  200000008: '_GMT_07_00_Chihuahua_LaPaz_Mazatlan',
  200000009: '_GMT_06_00_CentralAmerica',
  200000010: '_GMT_06_00_CentralTime_US_Canada_',
  200000011: '_GMT_06_00_Guadalajara_MexicoCity_Monterrey',
  200000012: '_GMT_06_00_Saskatchewan',
  200000013: '_GMT_05_00_Bogota_Lima_Quito_RioBranco',
  200000014: '_GMT_05_00_Chetumal',
  200000015: '_GMT_05_00_EasternTime_US_Canada_',
  200000016: '_GMT_05_00_Indiana_East_',
  200000017: '_GMT_04_30_Caracas',
  200000018: '_GMT_04_00_Asuncion',
  200000019: '_GMT_04_00_AtlanticTime_Canada_',
  200000020: '_GMT_04_00_LaPaz',
  200000021: '_GMT_04_00_Manaus',
  200000022: '_GMT_04_00_Santiago',
  200000023: '_GMT_03_30_Newfoundland',
  200000024: '_GMT_03_00_Brasilia',
  200000025: '_GMT_03_00_BuenosAires',
  200000026: '_GMT_03_00_BuenosAires_Georgetown',
  200000027: '_GMT_03_00_Greenland',
  200000028: '_GMT_03_00_Montevideo',
  200000029: '_GMT_03_00_Salvador',
  200000030: '_GMT_02_00_Mid_Atlantic',
  200000031: '_GMT_01_00_Azores',
  200000032: '_GMT_01_00_CaboVerdeIs_',
  200000033: '_GMT_Casablanca',
  200000034: '_GMT_CoordinatedUniversalTime',
  200000035: '_GMT_Casablanca_Monrovia_Reykjavik',
  200000036: '_GMT_GreenwichMeanTime_Dublin_Edinburgh_Lisbon_London',
  200000037: '_GMT_01_00_Amsterdam_Berlin_Bern_Rome_Stockholm_Vienna',
  200000038: '_GMT_01_00_Belgrade_Bratislava_Budapest_Ljubljana_Prague',
  200000039: '_GMT_01_00_Brussels_Copenhagen_Madrid_Paris',
  200000040: '_GMT_01_00_Sarajevo_Skopje_Warsaw_Zagreb',
  200000041: '_GMT_01_00_Tripoli',
  200000042: '_GMT_01_00_WestCentralAfrica',
  200000043: '_GMT_02_00_Amman',
  200000044: '_GMT_02_00_Athens_Bucharest_Istanbul',
  200000045: '_GMT_02_00_Beirut',
  200000046: '_GMT_02_00_Minsk',
  200000047: '_GMT_02_00_Cairo',
  200000048: '_GMT_02_00_Damascus',
  200000049: '_GMT_02_00_Harare_Pretoria',
  200000050: '_GMT_02_00_Helsinki_Kyiv_Riga_Sofia_Tallinn_Vilnius',
  200000051: '_GMT_02_00_Jerusalem',
  200000052: '_GMT_02_00_Windhoek',
  200000053: '_GMT_03_00_Istanbul',
  200000054: '_GMT_02_00_Kaliningrad_RTZ1_',
  200000055: '_GMT_03_00_Baghdad',
  200000056: '_GMT_03_00_Kuwait_Riyadh',
  200000057: '_GMT_03_00_Moscow_St_Petersburg_Volgograd_RTZ2_',
  200000058: '_GMT_04_00_PortLouis',
  200000059: '_GMT_03_00_Nairobi',
  200000060: '_GMT_03_00_Tbilisi',
  200000061: '_GMT_03_30_Tehran',
  200000062: '_GMT_04_00_AbuDhabi_Muscat',
  200000063: '_GMT_04_00_Baku',
  200000064: '_GMT_04_00_Izhevsk_Samara_RTZ3_',
  200000065: '_GMT_04_00_Yerevan',
  200000066: '_GMT_04_00_Yerevan_Donotuse_',
  200000067: '_GMT_04_30_Kabul',
  200000068: '_GMT_05_00_Ekaterinburg_RTZ4_',
  200000069: '_GMT_05_00_Islamabad_Karachi',
  200000070: '_GMT_05_00_Toshkent',
  200000071: '_GMT_05_30_Chennai_Kolkata_Mumbai_NewDelhi',
  200000072: '_GMT_05_30_SriJayawardenepura',
  200000073: '_GMT_05_45_Kathmandu',
  200000074: '_GMT_06_00_Nur_Sultan_Dhaka',
  200000075: '_GMT_06_00_Dhaka',
  200000076: '_GMT_06_00_Novosibirsk_RTZ5_',
  200000077: '_GMT_06_30_Yangon_Rangoon_',
  200000078: '_GMT_07_00_Bangkok_Hanoi_Jakarta',
  200000079: '_GMT_07_00_Krasnoyarsk_RTZ6_',
  200000080: '_GMT_08_00_Beijing_Chongqing_HongKong_Urumqi',
  200000081: '_GMT_08_00_Irkutsk_RTZ7_',
  200000082: '_GMT_08_00_Ulaanbaatar',
  200000083: '_GMT_08_00_KualaLumpur_Singapore',
  200000084: '_GMT_08_00_Perth',
  200000085: '_GMT_08_00_Taipei',
  200000086: '_GMT_09_00_Osaka_Sapporo_Tokyo',
  200000087: '_GMT_09_00_Seoul',
  200000088: '_GMT_09_00_Yakutsk_RTZ8_',
  200000089: '_GMT_09_30_Adelaide',
  200000090: '_GMT_09_30_Darwin',
  200000091: '_GMT_10_00_Brisbane',
  200000092: '_GMT_10_00_Canberra_Melbourne_Sydney',
  200000093: '_GMT_10_00_Guam_PortMoresby',
  200000094: '_GMT_10_00_Hobart',
  200000095: '_GMT_10_00_Vladivostok_Magadan_RTZ9_',
  200000096: '_GMT_11_00_Chokurdakh_RTZ10_',
  200000097: '_GMT_11_00_Magadan_SolomonIs__NewCaledonia',
  200000098: '_GMT_12_00_Magadan_Donotuse_',
  200000099: '_GMT_12_00_Anadyr_Petropavlovsk_Kamchatsky_RTZ11_',
  200000100: '_GMT_12_00_Auckland_Wellington',
  200000101: '_GMT_12_00_CoordinatedUniversalTime_12',
  200000102: '_GMT_12_00_Fiji_Kamchatka_MarshallIs_',
  200000103: '_GMT_13_00_Samoa',
  200000104: '_GMT_13_00_Nuku_alofa'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_deliveryaddresstimezone = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_deliveryaddresstimezone;
export const Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressorderspecific = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressorderspecific = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressorderspecific;
export const Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressprivate = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressprivate;
export const Mserp_purchaserequisitionlinev2entitiesmserp_ispartialdeliveryprevented = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_ispartialdeliveryprevented = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_ispartialdeliveryprevented;
export const Mserp_purchaserequisitionlinev2entitiesmserp_isprepaymentrequired = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_isprepaymentrequired = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_isprepaymentrequired;
export const Mserp_purchaserequisitionlinev2entitiesmserp_linestatus = {
  200000000: 'Draft',
  200000001: 'Inreview',
  200000002: 'Rejected',
  200000003: 'Approved',
  200000004: 'Cancelled',
  200000005: 'Closed',
  200000006: 'Budgetreserved'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_linestatus = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_linestatus;
export const Mserp_purchaserequisitionlinev2entitiesmserp_linetype = {
  200000000: 'Internalcatalogitem',
  200000001: 'Noncatalogitem',
  200000002: 'Externalcatalogitem'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_linetype = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_linetype;
export const Mserp_purchaserequisitionlinev2entitiesmserp_overridesalestax = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_overridesalestax = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_overridesalestax;
export const Mserp_purchaserequisitionlinev2entitiesmserp_rfqrequirementlevel = {
  200000000: 'None',
  200000001: 'Informal',
  200000002: 'Formal'
} as const;
export type Mserp_purchaserequisitionlinev2entitiesmserp_rfqrequirementlevel = keyof typeof Mserp_purchaserequisitionlinev2entitiesmserp_rfqrequirementlevel;

export interface Mserp_purchaserequisitionlinev2entitiesBase {
  mserp_accountingdate?: string;
  mserp_accountingdistributiontemplatename?: string;
  mserp_budgetreservationdocumentnumber?: string;
  mserp_budgetreservationlinenumber?: number;
  mserp_businessjustificationcode?: string;
  mserp_businessjustificationdetails?: string;
  mserp_buyinglegalentityid: string;
  mserp_currencycode?: string;
  mserp_defaultledgerdimensiondisplayvalue?: string;
  mserp_deliveryaddressbuildingcompliment?: string;
  mserp_deliveryaddresscity?: string;
  mserp_deliveryaddresscountryregionid?: string;
  mserp_deliveryaddresscountryregionisocode?: string;
  mserp_deliveryaddresscountyid?: string;
  mserp_deliveryaddressdescription?: string;
  mserp_deliveryaddressdistrictname?: string;
  mserp_deliveryaddressdunsnumber?: string;
  mserp_deliveryaddresslatitude?: number;
  mserp_deliveryaddresslocationid?: string;
  mserp_deliveryaddresslongitude?: number;
  mserp_deliveryaddressname?: string;
  mserp_deliveryaddresspostbox?: string;
  mserp_deliveryaddressstateid?: string;
  mserp_deliveryaddressstreet?: string;
  mserp_deliveryaddressstreetnumber?: string;
  mserp_deliveryaddresstimezone?: Mserp_purchaserequisitionlinev2entitiesmserp_deliveryaddresstimezone;
  mserp_deliveryaddresszipcode?: string;
  mserp_deliveryattentioninformation?: string;
  mserp_deliverycityinkana?: string;
  mserp_deliverystreetinkana?: string;
  mserp_externalitemnumber?: string;
  mserp_fixedassetreasoncode?: string;
  mserp_fixedassetrulequalifieroptionname?: string;
  mserp_fixedpricecharges?: number;
  "mserp_FK_PurchaseRequisitionHeaderV2_id@odata.bind"?: string;
  mserp_isdeliveryaddressorderspecific?: Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressorderspecific;
  mserp_isdeliveryaddressprivate?: Mserp_purchaserequisitionlinev2entitiesmserp_isdeliveryaddressprivate;
  mserp_ispartialdeliveryprevented?: Mserp_purchaserequisitionlinev2entitiesmserp_ispartialdeliveryprevented;
  mserp_isprepaymentrequired?: Mserp_purchaserequisitionlinev2entitiesmserp_isprepaymentrequired;
  mserp_itembatchnumber?: string;
  mserp_itemnumber?: string;
  mserp_lineamount?: number;
  mserp_linedescription?: string;
  mserp_linediscountamount?: number;
  mserp_linediscountpercentage?: number;
  mserp_linetype?: Mserp_purchaserequisitionlinev2entitiesmserp_linetype;
  mserp_overridesalestax?: Mserp_purchaserequisitionlinev2entitiesmserp_overridesalestax;
  mserp_prepaymentdetails?: string;
  mserp_primaryfield?: string;
  mserp_procurementproductcategoryname?: string;
  mserp_productcolorid?: string;
  mserp_productconfigurationid?: string;
  mserp_productname?: string;
  mserp_productsizeid?: string;
  mserp_productstyleid?: string;
  mserp_productversionid?: string;
  mserp_projectactivitynumber?: string;
  mserp_projectcategoryid?: string;
  mserp_projectid?: string;
  mserp_projectlinepropertyid?: string;
  mserp_projectsalescurrencycode?: string;
  mserp_projectsalesprice?: number;
  mserp_projectsalesunitsymbol?: string;
  mserp_projecttaxgroupcode?: string;
  mserp_projecttaxitemgroupcode?: string;
  mserp_purchaseprice?: number;
  mserp_purchasepricequantity?: number;
  mserp_purchaserequisitionlinev2entityid: string;
  mserp_purchaseunitsymbol?: string;
  mserp_receivingoperatingunitnumber?: string;
  mserp_receivingsiteid?: string;
  mserp_receivingwarehouseid?: string;
  mserp_requesteddate?: string;
  mserp_requestedpurchasequantity?: number;
  mserp_requisitionerpersonnelnumber?: string;
  mserp_requisitionlinenumber: number;
  mserp_requisitionnumber: string;
  mserp_salestaxgroupcode?: string;
  mserp_salestaxitemgroupcode?: string;
  mserp_url?: string;
  mserp_vendoraccountnumber?: string;
}

export interface Mserp_purchaserequisitionlinev2entities extends Mserp_purchaserequisitionlinev2entitiesBase {
  mserp_deliveryaddresstimezonename?: string;
  mserp_fixedassetgroupid?: string;
  mserp_fk_purchaserequisitionheaderv2_idname?: string;
  mserp_formatteddeliveryaddress?: string;
  mserp_isdeliveryaddressorderspecificname?: string;
  mserp_isdeliveryaddressprivatename?: string;
  mserp_ispartialdeliverypreventedname?: string;
  mserp_isprepaymentrequiredname?: string;
  mserp_linestatus?: Mserp_purchaserequisitionlinev2entitiesmserp_linestatus;
  mserp_linestatusname?: string;
  mserp_linetypename?: string;
  mserp_overridesalestaxname?: string;
  mserp_rfqrequirementlevel?: Mserp_purchaserequisitionlinev2entitiesmserp_rfqrequirementlevel;
  mserp_rfqrequirementlevelname?: string;
  mserp_fk_purchaserequisitionheaderv2_id?: object;
  _mserp_fk_purchaserequisitionheaderv2_id_value?: string;
}